+++ src/Analyser.php
@@ -72,34 +72,40 @@ namespace SebastianBergmann\PHPLOC
          * @var array
          */
         protected $count = array(
-          'files'              => 0,
-          'loc'                => 0,
-          'nclocClasses'       => 0,
-          'cloc'               => 0,
-          'ncloc'              => 0,
-          'ccn'                => 0,
-          'ccnMethods'         => 0,
-          'interfaces'         => 0,
-          'traits'             => 0,
-          'classes'            => 0,
-          'abstractClasses'    => 0,
-          'concreteClasses'    => 0,
-          'anonymousFunctions' => 0,
-          'functions'          => 0,
-          'methods'            => 0,
-          'publicMethods'      => 0,
-          'nonPublicMethods'   => 0,
-          'nonStaticMethods'   => 0,
-          'staticMethods'      => 0,
-          'constants'          => 0,
-          'classConstants'     => 0,
-          'globalConstants'    => 0,
-          'testClasses'        => 0,
-          'testMethods'        => 0,
-          'ccnByLoc'           => 0,
-          'ccnByNom'           => 0,
-          'nclocByNoc'         => 0,
-          'nclocByNom'         => 0
+          'files'                     => 0,
+          'loc'                       => 0,
+          'nclocClasses'              => 0,
+          'cloc'                      => 0,
+          'ncloc'                     => 0,
+          'ccn'                       => 0,
+          'ccnMethods'                => 0,
+          'interfaces'                => 0,
+          'traits'                    => 0,
+          'classes'                   => 0,
+          'abstractClasses'           => 0,
+          'concreteClasses'           => 0,
+          'anonymousFunctions'        => 0,
+          'functions'                 => 0,
+          'methods'                   => 0,
+          'publicMethods'             => 0,
+          'nonPublicMethods'          => 0,
+          'nonStaticMethods'          => 0,
+          'staticMethods'             => 0,
+          'constants'                 => 0,
+          'classConstants'            => 0,
+          'globalConstants'           => 0,
+          'testClasses'               => 0,
+          'testMethods'               => 0,
+          'ccnByLoc'                  => 0,
+          'ccnByNom'                  => 0,
+          'nclocByNoc'                => 0,
+          'nclocByNom'                => 0,
+          'methodCalls'               => 0,
+          'staticMethodCalls'         => 0,
+          'instanceMethodCalls'       => 0,
+          'attributeAccesses'         => 0,
+          'staticAttributeAccesses'   => 0,
+          'instanceAttributeAccesses' => 0
         );
 
         /**
@@ -174,16 +180,20 @@ namespace SebastianBergmann\PHPLOC
 
             $count = $this->count;
 
-            $count['directories']   = count($directories) - 1;
-            $count['namespaces']    = count($this->namespaces);
-            $count['classes']       = $count['abstractClasses'] +
-                                      $count['concreteClasses'];
-            $count['methods']       = $count['staticMethods'] +
-                                      $count['nonStaticMethods'];
-            $count['publicMethods'] = $count['methods'] -
-                                      $count['nonPublicMethods'];
-            $count['constants']     = $count['classConstants'] +
-                                      $count['globalConstants'];
+            $count['directories']       = count($directories) - 1;
+            $count['namespaces']        = count($this->namespaces);
+            $count['classes']           = $count['abstractClasses'] +
+                                          $count['concreteClasses'];
+            $count['methods']           = $count['staticMethods'] +
+                                          $count['nonStaticMethods'];
+            $count['publicMethods']     = $count['methods'] -
+                                          $count['nonPublicMethods'];
+            $count['constants']         = $count['classConstants'] +
+                                          $count['globalConstants'];
+            $count['attributeAccesses'] = $count['staticAttributeAccesses'] +
+                                          $count['instanceAttributeAccesses'];
+            $count['methodCalls']       = $count['staticMethodCalls'] +
+                                          $count['instanceMethodCalls'];
 
             if ($count['ncloc'] > 0) {
                 $count['ccnByLoc'] = $count['ccn'] / $count['ncloc'];
@@ -504,6 +514,40 @@ namespace SebastianBergmann\PHPLOC
                     case T_STRING: {
                         if ($value == 'define') {
                             $this->count['globalConstants']++;
+                        }
+                    }
+                    break;
+
+                    case T_DOUBLE_COLON:
+                    case T_OBJECT_OPERATOR: {
+                        $call = FALSE;
+                        $j    = $i + 1;
+
+                        while ($tokens[$j] != ';') {
+                            if (is_string($tokens[$j]) && $tokens[$j] == '(') {
+                                $call = TRUE;
+                            }
+
+                            else if (is_array($tokens[$j]) &&
+                                     $tokens[$j][0] == T_VARIABLE) {
+                                $access = TRUE;
+                            }
+
+                            $j++;
+                        }
+
+                        if ($call) {
+                            if ($token == T_DOUBLE_COLON) {
+                                $this->count['staticMethodCalls']++;
+                            } else {
+                                $this->count['instanceMethodCalls']++;
+                            }
+                        } else {
+                            if ($token == T_DOUBLE_COLON) {
+                                $this->count['staticAttributeAccesses']++;
+                            } else {
+                                $this->count['instanceAttributeAccesses']++;
+                            }
                         }
                     }
                     break;