+++ ChangeLog.md
@@ -6,6 +6,7 @@ This is the list of changes for the phploc 2.0 release series.
 phpcpd 2.0.0
 ------------
 
+* Implemented #19: Count number of static/non-static method calls (as well as accesses to static/non-static attributes).
 * Added the `--git-repository` switch to calculate the software metrics for each revision of a Git repository.
 * Added `--names-exclude` switch to exclude filenames using glob patterns.
 * Removed the ability to count the Executable Lines of Code (ELOC) and the optional dependency on Bytekit.